Move useRunKill and it's payload from RunPage to KillButton
Send Run query as KillButton props. Then do all
processing of payload and ownership in KillButton
component.

This also refreshes Run query after killing a run.
